organize types - remove unused dep
