+++ crates/telcoin-network-cli/src/db.rs
@@ -15,7 +15,7 @@ use std::{
 };
 use tn_config::TelcoinDirs as _;
 use tn_reth::{
-    iter_static_files, open_db_read_only, traits::TNPrimitives, DatabaseArguments, DatabaseEnv,
+    iter_static_files, open_db_read_only, TNPrimitives, DatabaseArguments, DatabaseEnv,
     RethDatabaseT as _, RethMdbxError, StaticFileProvider, Tables,
 };
 use tn_storage::{consensus_pack::DATA_NAME, pack_validate::validate_pack_file};
