+++ crates/tn-reth/Cargo.toml
@@ -32,7 +32,6 @@ reth-primitives = { workspace = true }
 reth-consensus = { workspace = true }
 reth-evm-ethereum = { workspace = true }
 reth-engine-primitives = { workspace = true }
-reth-engine-tree = { workspace = true }
 reth-revm = { workspace = true }
 reth-evm = { workspace = true }
 reth-node-builder = { workspace = true }
